% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/babel.R
\name{babel_transform}
\alias{babel_transform}
\title{Transform Code with Babel}
\usage{
babel_transform(code = "")
}
\arguments{
\item{code}{\code{character}}
}
\value{
transformed \code{character}
}
\description{
Helper function to use \code{V8} with \code{Babel} so we can
avoid a JSX transformer with using \code{reactR}.
}
\examples{
\dontrun{
library(reactR)
babel_transform('<div>react div</div>')
}
}

