% Generated by roxygen2 (4.0.2): do not edit by hand
\name{BindingContractObserved.S3}
\alias{BindingContractObserved.S3}
\title{Class: BindingContractObserved.S3}
\usage{
BindingContractObserved.S3(.x, where = .GlobalEnv, binding = function() { })
}
\arguments{
\item{.x}{\code{\link{ANY}}. An object of an arbitrary class whose class
   attribute should be updated so that it becomes an instance of class
   \code{BindingContractObserved.S3}. Mainly intended for rapid prototyping
   purposes}
}
\value{
Instance of class \code{BindingContractObserved.S3}.
}
\description{
Class representing git versions (S3) and its constructor function.
}
\section{Fields}{

\describe{
\item{\code{where}}{\code{\link{environment}}. Own environment.}

\item{\code{binding}}{\code{\link{function}}. Binding function.}
}}
\section{Intended use of this class}{


This S3 class, or to be more precise its constructor function, exists mainly
for rapid prototyping purposes.
This is mainly reflected in the fact, that when specifying \code{.x}, this
constructor function will simply update the \code{class} attribute of
whatever object has been provided.

However, it also allows for a more formal OOP-style of rapid
prototyping by offering explicit \emph{class fields} (all arguments except
\code{.x}). Nevertheless, it is probably advisable to switch to an
explicit formal approach such as \emph{S4} and/or \emph{Reference Classes}
once the package or application has reached a certain state of maturity.
}
\examples{
\dontrun{

## Informal use (intended mainly for rapid prototyping) //
## Takes *any* object and simply changes the class attributes
BindingContractObserved.S3(
  list(
    where = new.env(),
    binding = function(x) {x}
  )
)  
BindingContractObserved.S3(TRUE)  

## Formal use (explicitly using 'fields') //
BindingContractObserved.S3()
BindingContractObserved.S3(
  where = new.env(),
  binding = function(x) {x}
)

## Recommended: include namespace //
## Regardless if you plan on using this class in an informal or formal way
reactr::BindingContractObserved.S3(
  where = new.env(),
  binding = function(x) {x}
)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{BindingContractObserving.S3}
}
}

