% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{checkReactivityPrerequisites,ReactiveObject.S3-method}
\alias{checkReactivityPrerequisites,ReactiveObject.S3-method}
\title{Check Prerequisits for Setting Reactive Objects (ReactiveObject.S3-method)}
\usage{
\S4method{checkReactivityPrerequisites}{ReactiveObject.S3}(input = NULL,
  strict = 0, ...)
}
\arguments{
\item{input}{\code{\link{ReactiveObject.S3}}.}

\item{strict}{\code{\link{numeric}}.
Relevant if a referenced object has been removed thus breaking the
reactive relationship.
\itemiz{
  \item{\code{0}: } {function returns last cached value}
  \item{\code{1}: } {object value is set to \code{NULL} and is returned}
  \item{\code{2}: } {object value is set to an instance of condition class
      \code{BrokenReactiveBinding} and this condition is triggered whenever
      the object's value is requested by \code{\link[base]{get}} or
      its syntactical surgars \code{{obj-name} or \code{}}
  }
}}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{logical}}. \code{TRUE}: check passed;
   \code{FALSE}: check not passed.
}
\description{
See generic: \code{\link[reactr]{checkReactivityPrerequisites}}
}
\examples{
\dontrun{

## NOTE
## This function has a strong internal character as it is mainly designed to 
## be used inside of 'setReactiveS3()'.
## Thus not all aspects can be properly illustrated if not used in its default
## context

## Start with a fresh hash registry state //
resetRegistry()

suppressWarnings(rm(x_1))
input <- reactr::ReactiveObject.S3()
input <- prepareReactiveInstance(
  input = input, 
  id = "x_1",
  value = NULL,
  where = environment()
)

## Error reason: object does not exist yet //
try(checkReactivityPrerequisites(input = input, strict = 2))

## Error reason: object exists but is non-reactive //
x_1 <- 10
try(checkReactivityPrerequisites(input = input, strict = 2))

## Error reason: reactive object already exists //
## This cannot be illustrated outside the context of 'setReactiveS3()'

## All prerequisites met //
checkReactivityPrerequisites(input = input)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{checkReactivityPrerequisites},
    \link[reactr]{setReactiveS3}
}
}

