% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reacttools.R
\name{component}
\alias{component}
\title{Create a React component represented by an htmltools \code{\link[htmltools]{tag}}.}
\usage{
component(name, ...)
}
\arguments{
\item{name}{Name of the React component, which must start with an upper-case character.}

\item{...}{Attributes and children of the element to pass along to \code{\link[htmltools]{tag}} as varArgs.}
}
\description{
Create a React component represented by an htmltools \code{\link[htmltools]{tag}}.
}
