% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{copyReactive,character,missing,character,missing-method}
\alias{copyReactive,character,missing,character,missing-method}
\alias{copyReactive-character-missing-character-missing-method}
\title{Copy Reactive Object (character-missing-character-missing)}
\usage{
\S4method{copyReactive}{character,missing,character,missing}(id_from,
  where_from = parent.frame(), id_to, where_to = parent.frame(), ...)
}
\arguments{
\item{id_from}{\code{\link{character}}.}

\item{where_from}{\code{\link{missing}}.}

\item{id_to}{\code{\link{character}}.}

\item{where_to}{\strong{Signature argument}.
Object containing location information of object that the copy is
assigned to.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}

\item{where_from}{\code{\link{missing}}.}
}
\value{
See method
   \code{\link[reactr]{copyReactive-character-environment-character-environment-method}}
}
\description{
See generic: \code{\link[reactr]{copyReactive}}
}
\examples{
\dontrun{

##------------------------------------------------------------------------------
## setReactiveS3 --> ReactiveObject.S3 //
##------------------------------------------------------------------------------

setReactiveS3(id = "x_1", value = 10)
copyReactive(id_from = "x_1", id_to = "x_1_c")
x_1_c == x_1

x_1 <- 20
x_1_c
## --> x_1_c is not affected by the change in `x_1` as the two underlying 
## hidden objects are independent from each other:
identical(getFromRegistry("x_1"), getFromRegistry("x_1_c"))

## Classes //
class(getFromRegistry("x_1"))
class(getFromRegistry("x_1_c"))


##------------------------------------------------------------------------------
## setShinyReactive --> ReactiveShinyObject //
##------------------------------------------------------------------------------

setShinyReactive(id = "x_1", value = 10)
copyReactive(id_from = "x_1", id_to = "x_1_c")
x_1_c == x_1

x_1 <- 20
x_1_c
## --> x_1_c is not affected by the change in `x_1` as the two underlying 
## hidden objects are independent from each other:
identical(getFromRegistry("x_1"), getFromRegistry("x_1_c"))

## Classes //
class(getFromRegistry("x_1"))
class(getFromRegistry("x_1_c"))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{copyReactive},
   \link[reactr]{copyReactive-character-environment-character-environment-method}
}
}

