% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reacttools.R
\name{createReactInput}
\alias{createReactInput}
\title{Create a React-based input}
\usage{
createReactInput(inputId, class, dependencies, default = NULL,
  configuration = list(), container = htmltools::tags$div)
}
\arguments{
\item{inputId}{The \code{input} slot that will be used to access the value.}

\item{class}{Space-delimited list of CSS class names that should identify
this input type in the browser.}

\item{dependencies}{HTML dependencies to include in addition to those
supporting React. Must contain at least one dependency, that of the input's
implementation.}

\item{configuration}{Static configuration data.}

\item{container}{A function to generate an HTML element to contain the input.}

\item{value}{Initial value.}
}
\value{
A Shiny input suitable for inclusion in a UI.
}
\description{
Create a React-based input
}
