% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getBinding,character,missing-method}
\alias{getBinding,character,missing-method}
\alias{getBinding-character-missing-method}
\title{Get Binding Function (character-missing)}
\usage{
\S4method{getBinding}{character,missing}(id, where = parent.frame(), ...)
}
\arguments{
\item{id}{\code{\link{character}}.}

\item{where}{\code{\link{missing}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[reactr]{getBinding-character-environment-method}}
}
\description{
See generic: \code{\link[reactr]{getBinding}}
}
\examples{
\dontrun{

## Start with fresh registry //
resetRegistry()

## Reactive object w/o references //
setReactiveS3(id = "x_1", value = 10)
(binding <- getBinding(id = "x_1"))
identical(binding, getFromRegistry("x_1")$.func)

## Reactive object with references //
setReactiveS3(id = "x_2", value = function() "object-ref: {id: x_1}")
(binding <- getBinding(id = "x_2"))
identical(binding, getFromRegistry("x_2")$.func)

## Clean up //
rmReactive("x_1")
rmReactive("x_2")

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{getBinding},
   \link[reactr]{getBinding-character-environment-method}
}
}

