% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getBinding}
\alias{getBinding}
\title{Get Binding Function (generic)}
\usage{
getBinding(id, where = parent.frame(), ...)
}
\arguments{
\item{id}{\strong{Signature argument}.
Object containing name/ID information.}

\item{where}{\strong{Signature argument}.
Object containing location information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Retrieves the binding function by which a reactive object is bound to
other referenced reactive objects.
}
\details{
This is a convenience function to retrieve the value of field \code{.func}
from the invisible objects stored in the registry.
}
\examples{
\dontrun{

## Start with fresh registry //
resetRegistry()

## Reactive object w/o references //
setReactive(id = "x_1", value = 10)
(binding <- getBinding(id = "x_1"))
identical(binding, getFromRegistry("x_1")$.func)

## Reactive object with references //
setReactive(id = "x_2", value = function() "object-ref: {id: x_1}")
(binding <- getBinding(id = "x_2"))
identical(binding, getFromRegistry("x_2")$.func)

## Clean up //
rmReactive("x_1")
rmReactive("x_2")

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{getBinding-character-environment-method}
}
}

