% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getBoilerplateCode,BindingContractObserved.S3-method}
\alias{getBoilerplateCode,BindingContractObserved.S3-method}
\title{Get Boilerplate of Binding Contract}
\usage{
\S4method{getBoilerplateCode}{BindingContractObserved.S3}(ns, ...)
}
\arguments{
\item{ns}{\code{\link{BindingContractObserved.S3}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{call}}. Implemented binding interface.
}
\description{
See generic: \code{\link[reactr]{getBoilerplateCode}}
}
\examples{
\dontrun{

##------------------------------------------------------------------------------
## Important remarks //
##------------------------------------------------------------------------------
## This is a function that is currently only used inside a call to 'setReactive()'
## or 'setReactive_bare()'.
## Thus, in order to see what's going on here, we first need to mimick the 
## expected system state.
  
## Mimick expected system state //  
where <- new.env()

## Ensuring that 'test' exists:
setReactive(id = "test", value = Sys.time(), where = where)

## Getting boilerplate code for binding contract:
binding <- getBoilerplateCode(
  ns = reactr::BindingContractObserving.S3()
)

## We'd like to set a variable that monitors the variable 'test':
watch <- "test"

## We don't have to explicitly worry about variable that is monitoring 'test'
## except with respect to defining the binding relationship which I 
## call "binding contract".
## It defines the value of the variable monitoring 'test' based on the 
## value of 'test' via a function.
## This is the value of 'binding' when you call 'setReactive()'.
.binding <- function(x) {
  ## Add 24 hours //
  x + 60*60*24
}

## Evaluate binding contract //
## Note that it takes the value of 'where$test' and processes it 
## according to the binding relationship defined in '.binding()'
eval(binding)()

## Change monitored variable value //
where$test <- Sys.time()
where$test

## Re-evaluate binding contract //
eval(binding)()  

## All available boilerplate code of this package //
getBoilerplateCode(
  ns = reactr::BindingContractObserved.S3()
)
getBoilerplateCode(
  ns = reactr::BindingContractObserving.S3()
)
getBoilerplateCode(
  ns = reactr::BindingContractMutual.S3()
)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{getBoilerplateCode}
}
}

