% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getChecksum,character,environment-method}
\alias{getChecksum,character,environment-method}
\alias{getChecksum-character-environment-method}
\title{Get Checksum (character-environment)}
\usage{
\S4method{getChecksum}{character,environment}(id, where = parent.frame(), ...)
}
\arguments{
\item{id}{\code{\link{character}}.}

\item{where}{\code{\link{environment}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{function}}. Binding function.
}
\description{
See generic: \code{\link[reactr]{getChecksum}}
}
\examples{
\dontrun{

## Start with fresh registry //
resetRegistry()

## Reactive object w/o references //
setReactive(id = "x_1", value = 10)
getChecksum(id = "x_1")
## --> equivalent to:
digest::digest(10)

## Reactive object with references //
setReactive(id = "x_2", value = function() "object-ref: {id: x_1}")
getChecksum(id = "x_2")
## --> equivalent to:
digest::digest(x_1)

## Clean up //
rmReactive("x_1")
rmReactive("x_2")

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{getChecksum}
}
}

