% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getChecksum,character,missing-method}
\alias{getChecksum,character,missing-method}
\alias{getChecksum-character-missing-method}
\title{Get Checksum (character-missing)}
\usage{
\S4method{getChecksum}{character,missing}(id, where = parent.frame(), ...)
}
\arguments{
\item{id}{\code{\link{character}}.}

\item{where}{\code{\link{missing}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[reactr]{getChecksum-character-environment-method}}
}
\description{
See generic: \code{\link[reactr]{getChecksum}}
}
\examples{
\dontrun{

## Start with fresh registry //
resetRegistry()

## Reactive object w/o references //
setReactiveS3(id = "x_1", value = 10)
getChecksum(id = "x_1")
## --> equivalent to:
digest::digest(10)

## Reactive object with references //
setReactiveS3(id = "x_2", value = function() "object-ref: {id: x_1}")
getChecksum(id = "x_2")
## --> equivalent to:
digest::digest(x_1)

## Clean up //
rmReactive("x_1")
rmReactive("x_2")

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{getChecksum},
   \link[reactr]{getChecksum-character-environment-method}
}
}

