% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getChecksum}
\alias{getChecksum}
\title{Get Checksum (generic)}
\usage{
getChecksum(id, where = parent.frame(), ...)
}
\arguments{
\item{id}{\strong{Signature argument}.
Object containing name/ID information.}

\item{where}{\strong{Signature argument}.
Object containing location information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Retrieves the checksum of a reactive object's visible value.
}
\section{Main S4 method}{


\code{\link[reactr]{getChecksum-character-environment-method}}
}
\examples{
\dontrun{

## Start with fresh registry //
resetRegistry()

## Reactive object w/o references //
setReactive(id = "x_1", value = 10)
getChecksum(id = "x_1")
## --> equivalent to:
digest::digest(10)

## Reactive object with references //
setReactive(id = "x_2", value = function() "object-ref: {id: x_1}")
getChecksum(id = "x_2")
## --> equivalent to:
digest::digest(x_1)

## Clean up //
rmReactive("x_1")
rmReactive("x_2")

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{getChecksum-character-environment-method}
}
}

