% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFromRegistry}
\alias{getFromRegistry}
\title{Get From Registry (generic)}
\usage{
getFromRegistry(id, where = parent.frame(), ...)
}
\arguments{
\item{id}{\strong{Signature argument}.
Object containing name/ID information.}

\item{where}{\strong{Signature argument}.
Object containing location information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Retrieves reactive object from registry
(see \code{\link[reactr]{getRegistry}}).
}
\examples{
\dontrun{

## Example object //
resetRegistry()
setReactiveS3(id = "x_1", value = 10)
x_1

## Get from registry //
ls(getRegistry())
obj <- getFromRegistry(id = "x_1")
obj$id
obj$uid
obj$where

## Clean up //
obj$remove()
try(x_1)
## --> removes itself from 'where' and from registry
ls(getRegistry())

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{getFromRegistry-character-environment-method}
}
}

