% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getFromRegistryByUid,character-method}
\alias{getFromRegistry-character-method}
\alias{getFromRegistryByUid,character-method}
\title{Get From Registry By UID (character)}
\usage{
\S4method{getFromRegistryByUid}{character}(uid, ...)
}
\arguments{
\item{uid}{\code{\link{character}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}

\item{where}{\code{\link{environment}}.}
}
\value{
\code{\link{ReactiveObject.S3}}. Stored hidden instance of reactive
   object.
}
\description{
See generic: \code{\link[reactr]{getFromRegistryByUid}}
}
\examples{
\dontrun{

## Example object //
resetRegistry()
setReactiveS3(id = "x_1", value = 10)
x_1

## Get from registry //
uid <- computeObjectUid(id = "x_1")
showRegistry()
obj <- getFromRegistryByUid(uid = uid)
obj$.id
obj$.uid
obj$.where

## Clean up //
obj$.remove()
try(x_1)
## --> removes itself from 'where' and from registry
showRegistry()

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{getFromRegistryByUid}
}
}

