% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getFromRegistryByUid}
\alias{getFromRegistryByUid}
\title{Get From Registry By UID (generic)}
\usage{
getFromRegistryByUid(uid, ...)
}
\arguments{
\item{uid}{\strong{Signature argument}.
Object containing UID information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Retrieves reactive object from the registry
(see \code{\link[reactr]{getRegistry}}).
}
\examples{
\dontrun{

## Example object //
resetRegistry()
setReactiveS3(id = "x_1", value = 10)
x_1

## Get from registry //
uid <- getObjectUid(id = "x_1")
ls(getRegistry())
obj <- getFromRegistryByUid(uid = uid)
obj$id
obj$uid
obj$where

## Clean up //
obj$remove()
try(x_1)
## --> removes itself from 'where' and from registry
ls(getRegistry())

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{getFromRegistryByUid-character-environment-method}
}
}

