% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getHashRegistry,missing-method}
\alias{getHashRegistry,missing-method}
\title{Get Hash Registry (missing-method)}
\usage{
\S4method{getHashRegistry}{missing}(where = NULL, ...)
}
\arguments{
\item{where}{\code{\link{missing}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[reactr]{getHashRegistry-environment-method}}.
}
\description{
See generic: \code{\link[reactr]{getHashRegistry}}
}
\examples{
\dontrun{

## Set reactive so that there are elements in hash registry //  
resetHashRegistry()
setReactiveS3(id = "x_1", value = Sys.time())  
  
hash <- getHashRegistry()
hash
ls(hash)
identical(hash, getOption("reactr")$.hash)

## Clean up //
resetHashRegistry()

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{getHashRegistry-environment-method},
    \link[reactr]{removeFromHashRegistry}
}
}

