% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getHashRegistry}
\alias{getHashRegistry}
\title{Get Hash Registry}
\usage{
getHashRegistry(where = NULL, ...)
}
\arguments{
\item{where}{\strong{Signature argument}.
Object containing hash registry location information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Convenience function to retrieve the hash registry located in
\code{getOption("reactr")$.hash}.
}
\examples{
\dontrun{

## Set reactive so that there are elements in hash registry //  
resetHashRegistry()
setReactiveS3(id = "x_1", value = Sys.time())  
  
hash <- getHashRegistry()
hash
ls(hash)
identical(hash, getOption("reactr")$.hash)

## Clean up //
resetHashRegistry()

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{getHashRegistry-environment-method},
    \link[reactr]{removeFromHashRegistryByUid}
}
}

