% Generated by roxygen2 (4.0.2): do not edit by hand
\name{computeObjectUid}
\alias{computeObjectUid}
\title{Get Object UID (generic)}
\usage{
computeObjectUid(id, where = parent.frame(), ...)
}
\arguments{
\item{id}{\code{\link{character}}.
Name of the reactive object.}

\item{where}{\code{\link{environment}}.
Environment to create the object in.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{character}}. Object UID (checksum).
}
\description{
Computes the UID of an object based on its name/ID (\code{id}) and the
environment where it has been assigned to (\code{where}).
}
\examples{
\dontrun{

##------------------------------------------------------------------------------
## W/o explicit 'where' //
##------------------------------------------------------------------------------

id_1 <- "x_1"
id_2 <- "x_2"

setReactiveS3(id = id_1, value = 10)
setReactiveS3(id = id_2, value = function() {
  .ref_1 <- get(x = id_1)
})
x_1
x_2

uid_1 <- computeObjectUid(id = id_1)
uid_1
uid_2 <- computeObjectUid(id = id_2)
uid_2

## If these UIDs are available in the registry, they are correct //
registry <- getRegistry()
exists(uid_1, registry)
exists(uid_2, registry)

## Clean up //
rm(x_1)
rm(x_2)
resetRegistry()

##------------------------------------------------------------------------------
## With explicit 'where' //
##------------------------------------------------------------------------------

id_1 <- "x_1"
id_2 <- "x_2"
where <- new.env()

setReactiveS3(id = id_1, value = 10, where = where)
setReactiveS3(id = id_2, value = function() {
  .ref_1 <- get(x = id_1)
}, where = where)
where$x_1
where$x_2

uid_1 <- computeObjectUid(id = id_1, where = where)
uid_1
uid_2 <- computeObjectUid(id = id_2, where = where)
uid_2

## If these UIDs are available in the registry, they are correct //
registry <- getRegistry()
exists(uid_1, registry)
exists(uid_2, registry)

## Clean up //
rm(where)
resetRegistry()

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{setReactiveS3}
}
}

