% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getReactive,character,environment-method}
\alias{getReactive,character,environment-method}
\title{Get Reactive Object (character-environment)}
\usage{
\S4method{getReactive}{character,environment}(id, where = parent.frame(),
  hidden = FALSE, ...)
}
\arguments{
\item{id}{\code{\link{character}}.}

\item{where}{\code{\link{environment}}.}

\item{hidden}{\code{\link{logical}}.
\code{TRUE}: retrieve entire \strong{invisible} object/instance from registry;
\code{FALSE}: retrieve the visible part of the object/instance.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{ANY}}. Object/object value.
}
\description{
See generic: \code{\link[reactr]{getReactive}}
}
\examples{
\dontrun{ 

setReactiveS3(id = "x_1", value = 10)
getReactive(id = "x_1")
res <- getReactive(id = "x_1", hidden = TRUE)
res
class(res)

identical(res, getFromRegistry(id = "x_1"))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{getReactive}
}
}

