% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getReactive,character,missing-method}
\alias{getReactive,character,missing-method}
\title{Get Value from Environment}
\usage{
\S4method{getReactive}{character,missing}(id, where = .GlobalEnv,
  .hash_id = "._HASH", ...)
}
\arguments{
\item{id}{\code{\link{missing}}.}

\item{where}{\strong{Signature argument}.
Object containing environment information.}

\item{.hash_id}{\code{\link{character}}.
Name of the auxiliary environment for caching hash values.
Default: \code{"._HASH"}. Keep it unless this name is already taken in
either \code{where} or \code{where_watch}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[reactr]{getReactive-character-environment-method}}
}
\description{
See generic: \code{\link[reactr]{getReactive}}
}
\examples{
\dontrun{

where <- new.env()  
setReactive(id = "x_1", value = TRUE, where = where)
getReactive(id = "x_1", where = where)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{getReactive}
}
}

