% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getReactive}
\alias{getReactive}
\title{Get Value}
\usage{
getReactive(id, where = .GlobalEnv, .hash_id = "._HASH", ...)
}
\arguments{
\item{id}{\strong{Signature argument}.
Object containing path-like ID information.}

\item{where}{\strong{Signature argument}.
Object containing environment information.}

\item{.hash_id}{\code{\link{character}}.
Name of the auxiliary environment for caching hash values.
Default: \code{"._HASH"}. Keep it unless this name is already taken in
either \code{where} or \code{where_watch}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Retrieves value from an environment or any of its sub-environments based
on a path-like ID.
}
\details{
Values for \code{id} are expected to be of structure \code{a/b/c/.../z},
i.e. being a path-like identifier with a slash used as separator.
The identifier is transformed to \code{a$b$c$...$z} and then in turn to a
valid \emph{get} expression (\code{where$a$b$c$...$z}).
}
\examples{
\dontrun{

where <- new.env()  
setReactive(id = "x_1", value = TRUE, where = where)
getReactive(id = "x_1", where = where)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{getReactive-character-environment-method}
}
}

