% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getReactiveUid}
\alias{getReactiveUid}
\title{Get UID of Reactive Object}
\usage{
getReactiveUid(id, where = parent.frame(), ...)
}
\arguments{
\item{id}{\code{\link{character}}.
Name of the reactive object.}

\item{where}{\code{\link{environment}}.
Environment to create the object in.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Computes the UID of an reactive object.
}
\examples{
\dontrun{

## Set reactives so UIDs can be checked //
resetHashRegistry()

id_1 <- "x_1"
id_2 <- "x_2"

setReactiveS3(id = id_1, value = 10)
setReactiveS3(id = id_2, value = function() {
  .react_1 <- get(id_1, envir = where)
})

uid_1 <- getReactiveUid(id = id_1, where = where)
uid_1
uid_2 <- getReactiveUid(id = id_2, where = where)
uid_2

## If these UIDs are available in the hash registry, they are correct //
hash <- getHashRegistry()
exists(uid_1, hash)
exists(uid_2, hash)

## Clean up //
resetHashRegistry()

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{setReactiveS3}
}
}

