% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getValue,character,environment-method}
\alias{getValue,character,environment-method}
\title{Get Value from Environment}
\usage{
\S4method{getValue}{character,environment}(id, where = .GlobalEnv,
  .hash_id = "._HASH", ...)
}
\arguments{
\item{id}{\code{\link{character}}.}

\item{where}{\strong{Signature argument}.
Object containing environment information.}

\item{.hash_id}{\code{\link{character}}.
Name of the auxiliary environment for caching hash values.
Default: \code{"._HASH"}. Keep it unless this name is already taken in
either \code{where} or \code{where_watch}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{ANY}}. Variable value
}
\description{
See generic: \code{\link[reactr]{getValue}}
}
\examples{
\dontrun{

where <- new.env()  
setValue(id = "x_1", value = TRUE, where = where)
getValue(id = "x_1", where = where)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{getValue}
}
}

