% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{html_dependency_corejs}
\alias{html_dependency_corejs}
\title{Shim Dependency for React in RStudio Viewer}
\usage{
html_dependency_corejs()
}
\value{
\code{\link[htmltools]{htmlDependency}}
}
\description{
Add this first for 'React' to work in RStudio Viewer.
}
