% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{html_dependency_reacttools}
\alias{html_dependency_reacttools}
\title{Adds window.reactR.exposeComponents and window.reactR.hydrate}
\usage{
html_dependency_reacttools()
}
\description{
Adds window.reactR.exposeComponents and window.reactR.hydrate
}
