% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initializeHashRegistry,missing-method}
\alias{initializeHashRegistry,missing-method}
\title{Inititalize Hash Registry (missing-method)}
\usage{
\S4method{initializeHashRegistry}{missing}(where = NULL, ...)
}
\arguments{
\item{where}{\code{\link{missing}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[reactr]{initializeHashRegistry-NULL-method}}.
}
\description{
See generic: \code{\link[reactr]{initializeHashRegistry}}
}
\examples{
\dontrun{

## Make sure you remove the options that might have been loaded 
## at package startup //
options("reactr" = NULL)
getOption("reactr")

## Initialize hash registry //
initializeHashRegistry()
getOption("reactr")
getOption("reactr")$.hash
ls(getHashRegistry())

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{initializeHashRegistry-NULL-method},
    \link[reactr]{getHashRegistry}
}
}

