% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initializeHashRegistry}
\alias{initializeHashRegistry}
\title{Inititalize Hash Registry}
\usage{
initializeHashRegistry(where = NULL, ...)
}
\arguments{
\item{where}{\strong{Signature argument}.
Object containing hash registry location information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Convenience function to initialize the hash registry located in
\code{getOption("reactr")$.hash} in case it has not been properly
initialized at package startup.
}
\examples{
\dontrun{

## Make sure you remove the options that might have been loaded 
## at package startup //
options("reactr" = NULL)
getOption("reactr")

## Initialize hash registry //
initializeHashRegistry()
getOption("reactr")
getOption("reactr")$.hash
ls(getHashRegistry())

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{initializeHashRegistry-NULL-method},
    \link[reactr]{getHashRegistry}
}
}

