% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initializeOptionContainer,character-method}
\alias{initializeOptionContainer,character-method}
\title{Initialize Option Container (character-method)}
\usage{
\S4method{initializeOptionContainer}{character}(id, overwrite = FALSE, ...)
}
\arguments{
\item{id}{\code{\link{NULL}}.}

\item{overwrite}{\code{\link{logical}}.
\code{TRUE}: overwrite existing container;
\code{FALSE}: create only if no container exists yet.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{environment}}. The option container.
}
\description{
See generic: \code{\link[reactr]{initializeOptionContainer}}
}
\examples{
\dontrun{

options("reactr" = NULL)
getOption("reactr")
container <- initializeOptionContainer("reactr")
identical(getOption("reactr")$options, container$options)
identical(getOption("reactr")$.registry, container$.registry)  

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{initializeOptionContainer},
    \link[reactr]{getRegistry}
}
}

