% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initializeOptionContainer}
\alias{initializeOptionContainer}
\title{Initialize Option Container (generic)}
\usage{
initializeOptionContainer(id, overwrite = FALSE, ...)
}
\arguments{
\item{id}{\strong{Signature argument}.
Object containing R option ID information. Typically, this corresponds
to the name of the package/package project.}

\item{overwrite}{\code{\link{logical}}.
\code{TRUE}: overwrite existing container;
\code{FALSE}: create only if no container exists yet.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Convenience function to initialize the option container that is then stored
in option \code{"reactr"} and can be retrieved by
\code{getOption("reactr")}.
}
\examples{
\dontrun{

options("reactr" = NULL)
getOption("reactr")
container <- initializeOptionContainer("reactr")
identical(getOption("reactr")$options, container$options)
identical(getOption("reactr")$.registry, container$.registry)  

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{initializeOptionContainer-NULL-method},
    \link[reactr]{getRegistry}
}
}

