% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initializePackageOptions,character,environment-method}
\alias{initializePackageOptions,character,environment-method}
\alias{initializePackageOptions-character-environment-method}
\title{Initialize Package Options (character-environment)}
\usage{
\S4method{initializePackageOptions}{character,environment}(id = "options",
  where, ...)
}
\arguments{
\item{id}{\code{\link{character}}.}

\item{where}{\code{\link{environment}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{environment}}. Environment containing the options.
}
\description{
See generic: \code{\link[reactr]{initializePackageOptions}}
}
\examples{
\dontrun{

where <- new.env()
res <- initializePackageOptions(where = where)
exists(".registry", where, inherits = FALSE)
res <- initializePackageOptions(id = "test", where = where)
exists("test", where, inherits = FALSE)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{initializePackageOptions},
    \link[reactr]{getRegistry}
}
}

