% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initializePackageOptions,missing,missing-method}
\alias{initializePackageOptions,missing,missing-method}
\title{Initialize Package Options (missing-missing)}
\usage{
\S4method{initializePackageOptions}{missing,missing}(id = "options", where,
  ...)
}
\arguments{
\item{id}{\code{\link{missing}}.}

\item{where}{\code{\link{missing}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[reactr]{initializePackageOptions-NULL-method}}.
}
\description{
See generic: \code{\link[reactr]{initializePackageOptions}}
}
\examples{
\dontrun{

where <- new.env()
res <- initializePackageOptions(where = where)
exists(".registry", where, inherits = FALSE)
res <- initializePackageOptions(id = "test", where = where)
exists("test", where, inherits = FALSE)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{initializePackageOptions-NULL-method},
    \link[reactr]{getRegistry}
}
}

