% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initializeRegistry,character,environment-method}
\alias{initializeRegistry,character,environment-method}
\title{Initialize Registry (character-environment)}
\usage{
\S4method{initializeRegistry}{character,environment}(id = ".registry", where,
  ...)
}
\arguments{
\item{id}{\code{\link{character}}.}

\item{where}{\code{\link{environment}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{logical}}. \code{TRUE}.
}
\description{
See generic: \code{\link[reactr]{initializeRegistry}}
}
\examples{
\dontrun{

where <- new.env()
res <- initializeRegistry(where = where)
exists(".registry", where, inherits = FALSE)
res <- initializeRegistry(id = "test", where = where)
exists("test", where, inherits = FALSE)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{initializeRegistry},
    \link[reactr]{getRegistry}
}
}

