% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{initializeRegistry,missing,missing-method}
\alias{initializeRegistry,missing,missing-method}
\title{Initialize Registry (missing-missing)}
\usage{
\S4method{initializeRegistry}{missing,missing}(id = ".registry", where, ...)
}
\arguments{
\item{id}{\code{\link{missing}}.}

\item{where}{\code{\link{missing}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[reactr]{initializeRegistry-NULL-method}}.
}
\description{
See generic: \code{\link[reactr]{initializeRegistry}}
}
\examples{
\dontrun{

where <- new.env()
res <- initializeRegistry(where = where)
exists(".registry", where, inherits = FALSE)
res <- initializeRegistry(id = "test", where = where)
exists("test", where, inherits = FALSE)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{initializeRegistry-NULL-method},
    \link[reactr]{getRegistry}
}
}

