% Generated by roxygen2 (4.0.2): do not edit by hand
\name{initializeRegistry}
\alias{initializeRegistry}
\title{Initialize Registry (generic)}
\usage{
initializeRegistry(id = ".registry", where, ...)
}
\arguments{
\item{id}{\strong{Signature argument}.
Object containing ID information.}

\item{where}{\strong{Signature argument}.
Object containing location information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Convenience function to initialize a registry environment below an
option container.
}
\examples{
\dontrun{

where <- new.env()
res <- initializeRegistry(where = where)
exists(".registry", where, inherits = FALSE)
res <- initializeRegistry(id = "test", where = where)
exists("test", where, inherits = FALSE)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{initializeRegistry-NULL-method},
    \link[reactr]{getRegistry}
}
}

