% Generated by roxygen2 (4.0.2): do not edit by hand
\name{isReactive}
\alias{isReactive}
\title{Checks if Object is Reactive (generic)}
\usage{
isReactive(id, where = parent.frame(), ...)
}
\arguments{
\item{id}{\strong{Signature argument}.
Object containing name/ID information.}

\item{where}{\strong{Signature argument}.
Object containing location information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Checks if object is reactive by investigating if it

\itemize{
   \item{has an active binding via \code{\link[base]{bindingIsActive}}}
   \item{has reference of class \code{\link[reactr]{ReactiveObject.S3}}
     in registry (\code{\link[reactr]{getRegistry}})}
}
}
\examples{
\dontrun{

## Reactive objects //
setReactiveS3(id = "x_1", value = 10)
isReactive(id = "x_1")
unsetReactive("x_1")
isReactive(id = "x_1")

## Non-reactive/regular objects //
x_2 <- 10
isReactive(id = "x_2")
obj <- ReactiveObject.S3()
isReactive("obj")

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{isReactive-character-environment-method}
}
}

