% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react_provider.R
\name{plot_file}
\alias{plot_file}
\title{Title}
\usage{
plot_file(data)
}
\arguments{
\item{data}{a data frame constructed from csv file}
}
\value{
A plot generated with ggplot
}
\description{
Title
}
\examples{
\dontrun{
plot_file(mydata)
}
}

