% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.Reactive.S3}
\alias{print.Reactive.S3}
\title{Print Method}
\usage{
\method{print}{Reactive.S3}(x, as_is = TRUE, ...)
}
\arguments{
\item{x}{\code{\link[reactr]{Reactive.S3}}.
Instance to be printed.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Print method for class \code{Reactive.S3}
(see \code{\link[reactr]{Reactive.S3}}).
}
\examples{
\dontrun{

test <- reactr::Reactive.S3(
  value = 10,
  hash = digest::digest(10)
)
test
print(test)
print(test, as_is = FALSE)
res <- print(test, as_is = FALSE)
res
class(res)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{Reactive.S3}
}
}

