% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reacttools.R
\docType{data}
\name{React}
\alias{React}
\title{React component builder.}
\format{An object of class \code{react_component_builder} of length 0.}
\usage{
React
}
\description{
\code{React} is a syntactically-convenient way to create instances of React
components that can be sent to the browser for display. It is a list for
which \link[=InternalMethods]{extract methods} are defined, allowing
object creation syntax like \code{React$MyComponent(x = 1)} where
\code{MyComponent} is a React component you have exposed to Shiny in
JavaScript.
}
\details{
Internally, the \code{\link{component}} function is used to create the
component instance.
}
\examples{
# Create an instance of ParentComponent with two children,
# ChildComponent and OtherChildComponent.
React$ParentComponent(
  x = 1,
  y = 2,
  React$ChildComponent(),
  React$OtherChildComponent()
)
}
\keyword{datasets}
