% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reacttools.R
\name{reactData}
\alias{reactData}
\title{Create a data object for sending a React component to the client.}
\usage{
reactData(tag)
}
\arguments{
\item{tag}{React component or \code{\link[htmltools]{tag}}}
}
\description{
Create a data object for sending a React component to the client.
}
