% Generated by roxygen2 (4.0.2): do not edit by hand
\name{reactiveExpression}
\alias{reactiveExpression}
\title{Create Reactive Binding Object}
\usage{
reactiveExpression(x, env = parent.frame(), quoted = FALSE, label = NULL,
  domain = getDefaultReactiveDomain(), caller_offset = 1)
}
\arguments{
\item{x}{\code{\link{expression}} (quoted or unquoted).}

\item{env}{\code{\link{environment}}.
The parent environment for the reactive expression.
By default, this is the calling environment, the same as when defining
an ordinary non-reactive expression.}

\item{quoted}{\code{\link{logical}}.
\code{TRUE}: expression is quoted;
\code{FALSE}: expression is not quoted.
Default: \code{FALSE}. This is useful when you want to use an expression
that is stored in a variable; to do so, it must be quoted with \code{quote()}.}

\item{label}{\code{\link{character}}.
A label for the reactive expression, useful for debugging.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Creates a reactive object binding object. This is a slightly adapted form
of \code{\link[shiny]{reactive}} in order to make it compatible with being
called as an argument of \code{\link[reactr]{setShinyReactive}}.
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
    \link[shiny]{reactive}
}
}

