% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reacttools.R
\name{reactMarkup}
\alias{reactMarkup}
\title{Prepare data that represents a character vector, a React component, or an
htmltools tag for sending to the client.}
\usage{
reactMarkup(tag)
}
\arguments{
\item{tag}{character vector or React component or
\code{\link[htmltools]{tag}}}
}
\description{
Prepare data that represents a character vector, a React component, or an
htmltools tag for sending to the client.
}
