% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{reactr}
\alias{reactr}
\alias{reactr-package}
\title{Reactive object bindings with built-in caching}
\description{
Facilitates the specification of reactive object bindings based
on binding contracts. That way, a variable \code{x} can be dynamically
monitored by \code{n} other variables. Whenever \code{x} changes, the
\code{n} variables monitoring \code{x} change according to their binding
contracts that can be defined via regular functions (\code{\link{function}}).
}
\details{
The core functions/methods of this package:
 \itemize{
   \item{\code{\link[reactr]{setReactive}}: }{
     Facilitates the specification of reactive object bindings (S4 method).
   }
   \item{\code{\link[reactr]{setThis_bare}}: }{
     Bare S3 version of \code{\link[reactr]{setReactive}}. About roughly
     \code{10 - 15} faster than the S4 method.
   }
   \item{\code{\link[reactr]{getThis}}: }{
     Alternative way of retrieving variables from an environment.
   }
   \item{\code{\link[reactr]{getBoilerplateCode}}: }{
     Function that retrieves boilerplate code as required by
     \code{\link[reactr]{setReactive}} and \code{\link[base]{makeActiveBinding}}.
   }
}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}

