% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ReactrObservable}
\alias{ReactrObservable}
\title{Class: ReactrObservable}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x0000000009415fa0> 
 - attr(*, "name")= chr "ReactrObservable_generator"
}}
\usage{
ReactrObservable
}
\description{
Class that implements major parts of the reactivity mechanism of this
package.
}
\details{
Clone of \code{shiny:::Observable}. Only necessary because not all class
components of \code{shiny:::Observable} are exported.
}
\section{Fields}{


The fields correspond to the fields of \code{shiny:::Observable}.
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
    \link[reactr]{ReactiveShinyObject},
    \link[reactr]{setShinyReactive},
    \link[shiny]{reactive},
    \link[shiny]{makeReactiveBinding},
    \link[reactr]{setReactive}
}
}
\keyword{datasets}

