% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/react_provider.R
\name{read_file}
\alias{read_file}
\title{Title}
\usage{
read_file(file)
}
\arguments{
\item{file}{the name of a file starting from the package's working directory}
}
\value{
A data frame with the contents of file
}
\description{
Title
}
\examples{
\dontrun{
mydata <- read_file(csv_file)
}
}

