% Generated by roxygen2 (4.0.2): do not edit by hand
\name{removeFromRegistryByUid}
\alias{removeFromRegistryByUid}
\title{Remove From Registry}
\usage{
removeFromRegistryByUid(uid, ...)
}
\arguments{
\item{uid}{\strong{Signature argument}.
Object containing UID information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Removes entry of associated reactive object that has been unset via
\code{\link[reactr]{unsetReactiveByUid}} from the registry
\code{getOption("reactr")$.registry}.
}
\examples{
\dontrun{

## Set reactives so registry contains elements //  
id_1 <- "x_1"
id_2 <- "x_2"
where <- environment()

setReactiveS3(id = id_1, value = 10)
setReactiveS3(id = id_2, value = function() {
  .ref_1 <- get("x_1", envir = where)
})

## Comput UIDs //
uid_1 <- computeObjectUid(id = id_1, where = where)
uid_2 <- computeObjectUid(id = id_2, where = where)

## Inspect current state of registry //
registry <- getRegistry()
ls(registry)
exists(uid_1, registry)
exists(uid_2, registry)

## Remove //
removeFromRegistryByUid(uid = uid_1)
exists(uid_1, registry)
removeFromRegistryByUid(uid = uid_2)
exists(uid_2, registry)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{removeFromRegistryByUid-character-environment-method}
}
}

