% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{removeHashRegistry,missing-method}
\alias{removeHashRegistry,missing-method}
\title{Remove Hash Registry (missing-method)}
\usage{
\S4method{removeHashRegistry}{missing}(where = NULL, ...)
}
\arguments{
\item{where}{\code{\link{missing}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{logical}}. \code{TRUE}.
}
\description{
See generic: \code{\link[reactr]{removeHashRegistry}}
}
\examples{
\dontrun{

initializeHashRegistry()
getOption("reactr")$.hash
removeHashRegistry()
getOption("reactr")$.hash

## NOTE
## 'getHashRegistry()' calls 'initializeHashRegistry()' in order to ensure
## it, so it will never return 'NULL' as 'getOption("reactr")$.hash' would
getHashRegistry()

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{removeHashRegistry},
    \link[reactr]{getHashRegistry}
}
}

