% Generated by roxygen2 (4.0.2): do not edit by hand
\name{removeHashRegistry}
\alias{removeHashRegistry}
\title{Remove Hash Registry}
\usage{
removeHashRegistry(where = NULL, ...)
}
\arguments{
\item{where}{\strong{Signature argument}.
Object containing hash registry location information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Convenience function to remove the hash registry located in
\code{getOption("reactr")$.hash}.
}
\examples{
\dontrun{

initializeHashRegistry()
getOption("reactr")$.hash
removeHashRegistry()
getOption("reactr")$.hash

## NOTE
## 'getHashRegistry()' calls 'initializeHashRegistry()' in order to ensure
## it, so it will never return 'NULL' as 'getOption("reactr")$.hash' would
getHashRegistry()

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{removeHashRegistry-missing-method},
    \link[reactr]{removeFromHashRegistryByUid}
}
}

