% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{removeReactive,character,environment-method}
\alias{removeReactive,character,environment-method}
\alias{removeReactive-method_main}
\title{Remove Reactive Object (character-environment)}
\usage{
\S4method{removeReactive}{character,environment}(id, where = parent.frame(),
  ...)
}
\arguments{
\item{id}{\code{\link{character}}.}

\item{where}{\code{\link{environment}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{logical}}. \code{TRUE}: success; \code{FALSE}: object
   was not a reactive one or failure to remove.
}
\description{
See generic: \code{\link[reactr]{removeReactive}}
}
\examples{
\dontrun{

## Start with fresh registry //
resetRegistry()

## Non-strict //
where <- new.env()  
setReactiveS3(id = "x_1", value = 10, where = where)
setReactiveS3(id = "x_2", 
  value = function(refs = list(x_1 = where)) {x_1 * 2}, 
  where = where)

removeReactive(id = "x_1", where = where)
exists("x_1", envir = where, inherits = FALSE)
where$x_2

## Strict //
where <- new.env()  
setReactiveS3(id = "x_2", 
  value = function(refs = list(x_1 = where)) {x_1 * 2}, 
  where = where, 
  strict = TRUE
)
removeReactive(id = "x_1", where = where)
try(where$x_2)
## --> as referenced object 'x_1' does not exist anymore

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   Generic: \link[reactr]{removeReactive}
}
}

