% Generated by roxygen2 (4.0.2): do not edit by hand
\name{removeReactive}
\alias{removeReactive}
\title{Remove Reactive Object}
\usage{
removeReactive(id, where = .GlobalEnv, .hash_id = "._HASH",
  .tracelevel = 0, ...)
}
\arguments{
\item{id}{\strong{Signature argument}.
Object containing path-like ID information.}

\item{where}{\strong{Signature argument}.
Object containing location information.}

\item{.hash_id}{\code{\link{character}}.
Name of the auxiliary environment for caching hash values.
Default: \code{"._HASH"}. Keep it unless this name is already taken in
either \code{where} or \code{where_watch}.}

\item{.tracelevel}{\code{\link{numeric}}.
Verbosity level for tracing purposes. Value of \code{0} means
\emph{no tracing} whereas values of \code{> 0} can be used to fine
control tracing. The trace level can also be set as a global option when
using package \code{tracer} (\strong{not functional yet}).}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Removes a reactive object from its environment.
}
\details{
\strong{Note that tis is different from unsetting a reactive object
via \code{\link[reactr]{removeReactive}}}. It is equivalent to
\code{\link[base]{rm}} with a previous call to
\code{\link[reactr]{unsetReactive}}.
}
\note{
The main S4 method is
\code{\link[reactr]{removeReactive-character-environment-method}}.
}
\section{Implications with respect to observing variables}{

If other reactive variables have been observing the reactive variable that
has been removed, from this point on they will simply return the last value
that has been cached if \code{strict = FALSE} or \code{NULL} if
\code{strict = TRUE} when the observing object was set via
\code{\link[reactr]{setReactive}}
}
\examples{
\dontrun{

where = new.env()  

setReactive(id = "x_1", value = 10, where = where)
setReactive(id = "x_2", watch = "x_1", where = where)

removeReactive(id = "x_1", where = where)
exists("x_1", envir = where, inherits = FALSE)
where$x_2

## Strict //
where = new.env()  
  
setReactive(id = "x_1", value = 10, where = where)
setReactive(id = "x_2", watch = "x_1", where = where, strict = TRUE)
removeReactive(id = "x_1", where = where)
where$x_2
## --> as observed object 'x_1' does not exist anymore

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{removeReactive-character-environment-method}
}
}

