% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{removeReactiveByUid,character-method}
\alias{removeReactiveByUid,character-method}
\alias{removeReactiveByUid-method_main}
\title{Unset Reactive Object}
\usage{
\S4method{removeReactiveByUid}{character}(uid, ...)
}
\arguments{
\item{uid}{\code{\link{character}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[reactr]{removeReactiveByUid-character-method}}
}
\description{
See generic: \code{\link[reactr]{removeReactiveByUid}}
}
\examples{
\dontrun{

where = new.env()  
setReactiveS3(id = "x_1", value = 10, where = where)
setReactiveS3(id = "x_2", 
              value = function(refs = list(x_1 = where)) x_1,
              where = where)

unsetReactiveByUid(uid = getObjectUid("x_1", where))
where$x_1
identical(where$x_2, where$x_1)
where$x_1 <- 20
where$x_1
where$x_2
## --> no reactive binding to 'x_1' anymore

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   Generic: \link[reactr]{removeReactiveByUid}
}
}

