% Generated by roxygen2 (4.0.2): do not edit by hand
\name{removeReactiveByUid}
\alias{removeReactiveByUid}
\title{Remove Reactive Object (generic)}
\usage{
removeReactiveByUid(uid, ...)
}
\arguments{
\item{uid}{\strong{Signature argument}.
Object containing UID information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Removes a reactive object from its environment.
}
\details{
\strong{Note that tis is different from unsetting a reactive object
via \code{\link[reactr]{removeReactiveByUid}}}. It is equivalent to
\code{\link[base]{rm}} with a previous call to
\code{\link[reactr]{unsetReactiveByUid}}.
}
\note{
The main S4 method is
\code{\link[reactr]{removeReactiveByUid-character-method}}.
}
\section{Implications with respect to observing variables}{

If other reactive variables have been observing the reactive variable that
has been removed, from this point on they will simply return the last value
that has been cached if \code{strict = FALSE} or \code{NULL} if
\code{strict = TRUE} when the observing object was set via
\code{\link[reactr]{setReactiveS3}}
}
\examples{
\dontrun{

setReactiveS3(id = "x_1", value = 10)
isReactive("x_1")
setReactiveS3(id = "x_2", value = function() "object-ref: {id: x_1}")

unsetReactiveByUid(uid = computeObjectUid("x_1"))
isReactive("x_1")
## --> `x_1` is now a regular/non-reactive object again

## Actual values still identical //
x_1
identical(x_2, x_1)

## Implications for objects referencing `x_1` //
x_1 <- 20
x_1
x_2
## --> no reactive binding to `x_1` anymore

setReactiveS3(id = "x_2", value = function() "object-ref: {id: x_1}",
              strict_get = 1)
x_1 <- 10
x_2

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{removeReactiveByUid-character-method}
}
}

