% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{resetHashRegistry,environment-method}
\alias{resetHashRegistry,environment-method}
\title{Reset Hash Registry}
\usage{
\S4method{resetHashRegistry}{environment}(where = getHashRegistry(), ...)
}
\arguments{
\item{where}{\code{\link{environment}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{logical}}. \code{TRUE}: successful; \code{FALSE}: fail.
}
\description{
See generic: \code{\link[reactr]{resetHashRegistry}}
}
\examples{
\dontrun{

## Set reactives so hash registry has entries //  
setReactiveS3(id = "x_1", value = 10)
setReactiveS3(id = "x_2", value = function() {
  .ref_1 <- get("x_1", envir = where)
})

## Inspect current state of hash registry //
hash <- getHashRegistry()
ls(hash)

## Reset //
resetHashRegistry()
ls(hash)
## --> empty 

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{resetHashRegistry},
    \link[reactr]{removeFromHashRegistryByUid}
}
}

