% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{resetHashRegistry,missing-method}
\alias{resetHashRegistry,missing-method}
\title{Reset Hash Registry}
\usage{
\S4method{resetHashRegistry}{missing}(where = getHashRegistry(), ...)
}
\arguments{
\item{where}{\code{\link{missing}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[reactr]{resetHashRegistry-environment-method}}.
}
\description{
See generic: \code{\link[reactr]{resetHashRegistry}}
}
\examples{
\dontrun{

## Set reactives so hash registry has entries //  
setReactiveS3(id = "x_1", value = 10)
setReactiveS3(id = "x_2", value = function() {
  .react_1 <- get("x_1", envir = where)
})

## Inspect current state of hash registry //
hash <- getHashRegistry()
ls(hash)

## Reset //
resetHashRegistry()
ls(hash)
## --> empty 

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{resetHashRegistry-environment-method},
    \link[reactr]{removeFromHashRegistry}
}
}

