% Generated by roxygen2 (4.0.2): do not edit by hand
\name{resetHashRegistry}
\alias{resetHashRegistry}
\title{Reset Hash Registry}
\usage{
resetHashRegistry(where = getHashRegistry(), ...)
}
\arguments{
\item{where}{\strong{Signature argument}.
Object containing hash registry location information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Resets the hash registry located in \code{getOption("reactr")$.hash}.
}
\examples{
\dontrun{

## Set reactives so hash registry has entries //  
setReactiveS3(id = "x_1", value = 10)
setReactiveS3(id = "x_2", value = function() {
  .react_1 <- get("x_1", envir = where)
})

## Inspect current state of hash registry //
hash <- getHashRegistry()
ls(hash)

## Reset //
resetHashRegistry()
ls(hash)
## --> empty 

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{resetHashRegistry-environment-method},
    \link[reactr]{removeFromHashRegistry}
}
}

