% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{resetRegistry,missing-method}
\alias{resetRegistry,missing-method}
\alias{resetRegistry-missing-method}
\title{Reset Registry (missing)}
\usage{
\S4method{resetRegistry}{missing}(where = getRegistry(), ...)
}
\arguments{
\item{where}{\code{\link{missing}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[reactr]{resetRegistry-environment-method}}.
}
\description{
See generic: \code{\link[reactr]{resetRegistry}}
}
\examples{
\dontrun{

## Set reactives so registry has entries //  
setReactiveS3(id = "x_1", value = 10)
setReactiveS3(id = "x_2", value = function() {
  .ref_1 <- get("x_1", envir = where)
})

## Inspect current state of registry //
registry <- getRegistry()
ls(registry)

## Reset //
resetRegistry()
ls(registry)
## --> empty 

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{resetRegistry-environment-method},
    \link[reactr]{removeFromRegistryByUid}
}
}

