% Generated by roxygen2 (4.0.2): do not edit by hand
\name{resetRegistry}
\alias{resetRegistry}
\title{Reset Registry}
\usage{
resetRegistry(where = getRegistry(), ...)
}
\arguments{
\item{where}{\strong{Signature argument}.
Object containing hash registry location information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Resets the hash registry located in \code{getOption("reactr")$.registry}.
}
\examples{
\dontrun{

## Set reactives so hash registry has entries //  
setReactiveS3(id = "x_1", value = 10)
setReactiveS3(id = "x_2", value = function() {
  .ref_1 <- get("x_1", envir = where)
})

## Inspect current state of hash registry //
hash <- getRegistry()
ls(hash)

## Reset //
resetRegistry()
ls(hash)
## --> empty 

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{resetRegistry-environment-method},
    \link[reactr]{removeFromRegistryByUid}
}
}

