% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold_input.R
\name{scaffoldReactShinyInput}
\alias{scaffoldReactShinyInput}
\title{Create implementation scaffolding for a React.js-based Shiny input.}
\usage{
scaffoldReactShinyInput(name, npmPkgs = NULL, edit = interactive())
}
\arguments{
\item{name}{Name of input}

\item{npmPkgs}{Optional \href{https://npmjs.com/}{NPM} packages upon which
this input is based which will be used to populate \code{package.json}.
Should be a named list of names to
\href{https://docs.npmjs.com/files/package.json#dependencies}{versions}.}

\item{edit}{Automatically open the input's source files after creating the
scaffolding.}
}
\description{
Add the minimal code required to implement a React.js-based Shiny input to an
R package.
}
\note{
This function must be executed from the root directory of the package
  you wish to add the input to.
}
