% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{setReactive,character,ANY,missing,character,missing,missing-method}
\alias{setReactive,character,ANY,missing,character,missing,missing-method}
\title{Set Value (character,ANY,missing,character,missing,missing)}
\usage{
\S4method{setReactive}{character,ANY,missing,character,missing,missing}(id,
  value = NULL, where = .GlobalEnv, watch = character(),
  where_watch = where, binding = substitute(expression()),
  binding_type = 1, mutual = FALSE, force = FALSE, strict = FALSE,
  .hash_id = "._HASH", .tracelevel = 0, ...)
}
\arguments{
\item{id}{\code{\link{character}}.}

\item{value}{\code{\link{ANY}}.}

\item{where}{\code{\link{missing}}.}

\item{watch}{\code{\link{character}}.}

\item{where_watch}{\code{\link{missing}}.}

\item{binding}{\code{\link{missing}}.}

\item{binding_type}{\code{\link{integer}} or \code{\link{numerical}}.
\itemize{
   \item{\code{1}:} {
   use \code{\link{makeActiveBinding}} to establish the binding
   }
   \item{\code{2}:} {
   use custom way of establishing the binding
   }
}}

\item{mutual}{\code{\link{logical}}.
\code{TRUE}: mutual binding contract;
\code{FALSE}: binding contract depends on \code{watch} being specified
or not (specified: \code{monitoring} contract;
not specified: \emph{monitored} contract).}

\item{force}{\code{\link{logical}}.
\code{TRUE}: force a binding reset even though there might already
have been defined another one;
\code{FALSE}: in case a binding has already been defined it is not
overwritten. \strong{Note that for the following constellations this value is
automtically set to \code{TRUE}: \code{mutual = TRUE} and whenever an
explicit binding definition is provided via \code{binding}}.}

\item{strict}{\code{\link{logical}}.
\code{TRUE}: if object is removed, all observing objects are set to
\code{NULL} as returning the last cached value would be
misleading (that would be the case if the object has just been
unset by \code{\link[reactr]{unsetReactive}});
\code{FALSE}: observing object return the last cached value of the
linked to the removed object.}

\item{.hash_id}{\code{\link{character}}.
Name of the auxiliary environment for caching hash values.
Default: \code{"._HASH"}. Keep it unless this name is already taken in
either \code{where} or \code{where_watch}.}

\item{.tracelevel}{\code{\link{numeric}}.
Verbosity level for tracing purposes. Value of \code{0} means
\emph{no tracing} whereas values of \code{> 0} can be used to fine
control tracing. The trace level can also be set as a global option when
using package \code{tracer} (\strong{not functional yet}).}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[reactr]{setReactive-character-ANY-missing-character-missing-call-method}}.
}
\description{
See generic: \code{\link[reactr]{setReactive}}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   Generic: \link[reactr]{setReactive}
}
}

