% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setReactive_reactive,character,ANY,environment,character,environment,call-method}
\alias{setReactive_reactive,character,ANY,environment,character,environment,call-method}
\alias{setReactive_reactive-method_main}
\title{Set Value (character,ANY,environment,character,environment,call)}
\usage{

  \S4method{setReactive_reactive}{character,ANY,environment,character,environment,call}(id,
  value, where, watch, where_watch, binding, binding_type, mutual, force,
  .hash_id, .tracelevel, .binding = NULL, ...)
}
\arguments{
\item{id}{\code{\link{character}}.}

\item{value}{\code{\link{ANY}}.}

\item{where}{\code{\link{environment}}.}

\item{watch}{\code{\link{character}}.}

\item{where_watch}{\code{\link{environment}}.}

\item{binding}{\code{\link{call}}.}

\item{.binding}{\code{\link{function}}.
Internal argument that should not be set explicitly.
The value at runtime will correspond to the function that has been
provided via argument \code{binding}.}

\item{binding_type}{\code{\link{integer}} or \code{\link{numerical}}.
\itemize{
   \item{\code{1}:} {
   use \code{\link{makeActiveBinding}} to establish the binding
   }
   \item{\code{2}:} {
   use custom way of establishing the binding
   }
}}

\item{mutual}{\code{\link{logical}}.
\code{TRUE}: mutual binding contract;
\code{FALSE}: binding contract depends on \code{watch} being specified
or not (specified: \code{monitoring} contract;
not specified: \emph{monitored} contract).}

\item{force}{\code{\link{logical}}.
\code{TRUE}: force a binding reset even though there might already
have been defined another one;
\code{FALSE}: in case a binding has already been defined it is not
overwritten. \strong{Note that for the following constellations this value is
automtically set to \code{TRUE}: \code{mutual = TRUE} and whenever an
explicit binding definition is provided via \code{binding}}.}

\item{.hash_id}{\code{\link{character}}.
Name of the auxiliary environment for caching hash values.
Default: \code{"._HASH"}. Keep it unless this name is already taken in
either \code{where} or \code{where_watch}.}

\item{.tracelevel}{\code{\link{numeric}}.
Verbosity level for tracing purposes. Value of \code{0} means
\emph{no tracing} whereas values of \code{> 0} can be used to fine
control tracing. The trace level can also be set as a global option when
using package \code{tracer} (\strong{not functional yet}).}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{ANY}}. Value of \code{value} or the return value
   of the binding function.
}
\description{
See generic: \code{\link[reactr]{setReactive_reactive}}
}
\examples{
\dontrun{

## TODO: add example

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   Generic: \link[reactr]{setReactive_reactive}
}
}

