% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setReactive_reactive,character,ANY,environment,character,missing,missing-method}
\alias{setReactive_reactive,character,ANY,environment,character,missing,missing-method}
\title{Set Value (character,ANY,environment,character,missing,missing)}
\usage{

  \S4method{setReactive_reactive}{character,ANY,environment,character,missing,missing}(id,
  value = NULL, where = .GlobalEnv, watch = character(),
  where_watch = where, binding = substitute(expression()),
  binding_type = 1, mutual = FALSE, force = FALSE, .hash_id = "._HASH",
  .tracelevel = 0, ...)
}
\arguments{
\item{id}{\code{\link{character}}.}

\item{value}{\code{\link{ANY}}.}

\item{where}{\code{\link{environment}}.}

\item{watch}{\code{\link{character}}.}

\item{where_watch}{\code{\link{missing}}.}

\item{binding}{\code{\link{missing}}.}

\item{binding_type}{\code{\link{integer}} or \code{\link{numerical}}.
\itemize{
   \item{\code{1}:} {
   use \code{\link{makeActiveBinding}} to establish the binding
   }
   \item{\code{2}:} {
   use custom way of establishing the binding
   }
}}

\item{mutual}{\code{\link{logical}}.
\code{TRUE}: mutual binding contract;
\code{FALSE}: binding contract depends on \code{watch} being specified
or not (specified: \code{monitoring} contract;
not specified: \emph{monitored} contract).}

\item{force}{\code{\link{logical}}.
\code{TRUE}: force a binding reset even though there might already
have been defined another one;
\code{FALSE}: in case a binding has already been defined it is not
overwritten. \strong{Note that for the following constellations this value is
automtically set to \code{TRUE}: \code{mutual = TRUE} and whenever an
explicit binding definition is provided via \code{binding}}.}

\item{.hash_id}{\code{\link{character}}.
Name of the auxiliary environment for caching hash values.
Default: \code{"._HASH"}. Keep it unless this name is already taken in
either \code{where} or \code{where_watch}.}

\item{.tracelevel}{\code{\link{numeric}}.
Verbosity level for tracing purposes. Value of \code{0} means
\emph{no tracing} whereas values of \code{> 0} can be used to fine
control tracing. The trace level can also be set as a global option when
using package \code{tracer} (\strong{not functional yet}).}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[reactr]{setReactive_reactive-character-ANY-environment-character-missing-call-method}}.
}
\description{
See generic: \code{\link[reactr]{setReactive_reactive}}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   Generic: \link[reactr]{setReactive_reactive}
}
}

