% Generated by roxygen2 (4.0.2): do not edit by hand
\name{setReactiveSource}
\alias{setReactiveSource}
\title{Create Reactive Source Object}
\usage{
setReactiveSource(id, value = NULL, where = parent.frame(),
  overwrite = TRUE, typed = FALSE)
}
\arguments{
\item{id}{\code{\link{character}}.
Name/ID of the reactive source object.}

\item{value}{\code{\link{ANY}}.
Value of reactive source object.}

\item{where}{\code{\link{environment}}.
Environment in which to create the object.}

\item{overwrite}{\code{\link{logical}}.
Only relevant if object already exists.
\code{TRUE}: overwrite existing value;
\code{FALSE}: keep currrent value.}

\item{typed}{\code{\link{logical}}.
\code{TRUE}: checks class validity of assignment value specified via
\code{value} and throws an error if classes do not match or if the class
of the assignment value does not inherit from the class of field value
\code{.value} at initialization;
\code{FALSE}: no class check is performed.
Note that initial values of \code{NULL} are disregarded for the class
check, i.e. any value overwriting an initial \code{NULL} value
is valid.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Creates a reactive source object that can be used by
observable reactive objects created via \code{\link[reactr]{setShinyReactive}}
or \code{\link[shiny]{reactive}}.
}
\details{
This is a slightly modified version of \link[shiny]{makeReactiveBinding}.
}
\examples{
\dontrun{

## Create reactive source //
setReactiveSource(id = "x_1", value = 10)
x_1
x_1 <- 20
x_1

## No overwrite //
setReactiveSource(id = "x_1", value = 10)
x_1
setReactiveSource(id = "x_1", value = 20, overwrite = FALSE)
x_1
## --> as `x_1` already existed with a non-NULL value and `overwrite = FALSE`,
## no overwrite has been performed

## Typed //
setReactiveSource(id = "x_1", value = 10, typed = TRUE)
x_1 <- 20
try(x_1 <- "hello world!")

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{setShinyReactive},
   \link[shiny]{reactive}
}
}

