% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setValue,character,ANY,environment,character,environment,missing-method}
\alias{setValue,character,ANY,environment,character,environment,missing-method}
\title{Set Value (character,ANY,environment,character,environment,missing)}
\usage{

  \S4method{setValue}{character,ANY,environment,character,environment,missing}(id,
  value = NULL, where = .GlobalEnv, watch = character(),
  where_watch = where, binding = substitute(expression()),
  binding_type = 1, mutual = FALSE, .hash_id = "._HASH",
  .tracelevel = 0, ...)
}
\arguments{
\item{id}{\code{\link{character}}.}

\item{value}{\code{\link{ANY}}.}

\item{where}{\code{\link{environment}}.}

\item{watch}{\code{\link{character}}.}

\item{where_watch}{\code{\link{environment}}.}

\item{binding}{\code{\link{missing}}.}

\item{binding_type}{\code{\link{integer}} or \code{\link{numerical}}.
\itemize{
   \item{\code{1}:} {
   use \code{\link{makeActiveBinding}} to establish the binding
   }
   \item{\code{2}:} {
   use custom way of establishing the binding
   }
}}

\item{mutual}{\code{\link{logical}}.
\code{TRUE}: mutual binding contract;
\code{FALSE}: binding contract depends on \code{watch} being specified
or not (specified: \code{monitoring} contract;
not specified: \emph{monitored} contract)}

\item{.hash_id}{\code{\link{character}}.
Name of the auxiliary environment for caching hash values.
Default: \code{"._HASH"}. Keep it unless this name is already taken in
either \code{where} or \code{where_watch}.}

\item{.tracelevel}{\code{\link{numeric}}.
Verbosity level for tracing purposes. Value of \code{0} means
\emph{no tracing} whereas values of \code{> 0} can be used to fine
control tracing. The trace level can also be set as a global option when
using package \code{tracer} (\strong{not functional yet}).}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[reactr]{setValue-character-ANY-environment-character-environment-call-method}}.
}
\description{
See generic: \code{\link[reactr]{setValue}}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   Generic: \link[reactr]{setValue}
}
}

