% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{showPushRefs,character,environment-method}
\alias{showPushRefs,character,environment-method}
\alias{showPushRefs-environment-method}
\title{Show Push References (environment)}
\usage{
\S4method{showPushRefs}{character,environment}(id, where = parent.frame(),
  ...)
}
\arguments{
\item{id}{\code{\link{character}}.}

\item{where}{\code{\link{environment}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{logical}}. \code{TRUE}: successful; \code{FALSE}: fail.
}
\description{
See generic: \code{\link[reactr]{showPushRefs}}
}
\examples{
\dontrun{

setReactive(id = "x_1", value = 10)
setReactive(id = "x_2", value = function() "object-ref: {id: x_1}")

showPushRefs(id = "x_1")
showPushRefs(id = "x_2")

setReactive(id = "x_2", value = function() "object-ref: {id: x_1}", push = TRUE)
showPushRefs(id = "x_2")

## Clean up //
rmReactive("x_1")
rmReactive("x_2")

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{showPushRefs},
    \link[reactr]{getRegistry}
}
}

