% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{showPushRefs,character,missing-method}
\alias{showPushRefs,character,missing-method}
\alias{showPushRefs-missing-method}
\title{Show Push References (missing)}
\usage{
\S4method{showPushRefs}{character,missing}(id, where = parent.frame(), ...)
}
\arguments{
\item{id}{\code{\link{character}}.}

\item{where}{\code{\link{missing}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[reactr]{showPushRefs-character-environment-method}}.
}
\description{
See generic: \code{\link[reactr]{showPushRefs}}
}
\examples{
\dontrun{

setReactiveS3(id = "x_1", value = 10)
setReactiveS3(id = "x_2", value = function() "object-ref: {id: x_1}")

showPushRefs(id = "x_1")
showPushRefs(id = "x_2")

setReactiveS3(id = "x_2", value = function() "object-ref: {id: x_1}", push = TRUE)
showPushRefs(id = "x_2")

## Clean up //
rmReactive("x_1")
rmReactive("x_2")

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{showPushRefs-environment-method}
}
}

