% Generated by roxygen2 (4.0.2): do not edit by hand
\name{showPushRefs}
\alias{showPushRefs}
\title{Show Push References (generic)}
\usage{
showPushRefs(id, where = parent.frame(), ...)
}
\arguments{
\item{id}{\strong{Signature argument}.
Object containing name/ID information.}

\item{where}{\strong{Signature argument}.
Object containing location information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Convenience function to show the UIDs of registered push references.
}
\section{Main S4 method}{


\code{\link[reactr]{showPushRefs-character-environment-method}}
}
\examples{
\dontrun{

setReactive(id = "x_1", value = 10)
setReactive(id = "x_2", value = function() "object-ref: {id: x_1}")

showPushRefs(id = "x_1")
showPushRefs(id = "x_2")

setReactive(id = "x_2", value = function() "object-ref: {id: x_1}", push = TRUE)
showPushRefs(id = "x_2")

## Clean up //
rmReactive("x_1")
rmReactive("x_2")

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{showPushRefs-environment-method}
}
}

