% Generated by roxygen2 (4.0.1): do not edit by hand
\name{unsetReactive}
\alias{unsetReactive}
\title{Unset Reactive Object}
\usage{
unsetReactive(id, where = .GlobalEnv, .hash_id = "._HASH",
  .tracelevel = 0, ...)
}
\arguments{
\item{id}{\strong{Signature argument}.
Object containing path-like ID information.}

\item{where}{\strong{Signature argument}.
Object containing location information.}

\item{.hash_id}{\code{\link{character}}.
Name of the auxiliary environment for caching hash values.
Default: \code{"._HASH"}. Keep it unless this name is already taken in
either \code{where} or \code{where_watch}.}

\item{.tracelevel}{\code{\link{numeric}}.
Verbosity level for tracing purposes. Value of \code{0} means
\emph{no tracing} whereas values of \code{> 0} can be used to fine
control tracing. The trace level can also be set as a global option when
using package \code{tracer} (\strong{not functional yet}).}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Removes the reactive binding from an object
See main method
\code{\link{unsetReactive-character-environment-method}}
}
\note{
The main S4 method is
\code{\link[reactr]{unsetReactive-character-environment-method}}.
}
\section{Implications with respect to observing variables}{

If other reactive variables have been observing the reactive variable that
has been unset, from this point on they will simply return the last value
that has been cached.
}
\examples{
\dontrun{

where <- new.env()  

setReactive(id = "x_1", value = 10, where = where)
setReactive(id = "x_2", watch = "x_1", where = where)

## Illustrate reactiveness //
where$x_1
where$x_2
where$x_1 <- 50
where$x_1 
where$x_2

## Unset reactive --> turn it into a regular object again //
unsetReactive(id = "x_1", where = where)

## Illustrate removed reactiveness //
where$x_1
where$x_2
where$x_1 <- 10
where$x_1
where$x_2
## --> 'x_1' is not a reactive object anymore; from now on, 'x_2' simply returns
## the last value that has been cached

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{unsetReactive-character-environment-method}
}
}

