% Generated by roxygen2 (4.0.2): do not edit by hand
\name{checkReactivityPrerequisites}
\alias{checkReactivityPrerequisites}
\title{Check Prerequisits for Setting Reactive Objects (generic)}
\usage{
checkReactivityPrerequisites(input = NULL, strict = 0, ...)
}
\arguments{
\item{input}{\strong{Signature argument}.
Object containing a suitable object for the check.
Typically, this corresponds to the instance of class
\code{\link[reactr]{ReactiveObject.S3}}.}

\item{strict}{\code{\link{numeric}}.
Relevant when initially setting a reactive object
\itemize{
  \item{\code{0}: } {no checks are performed}
  \item{\code{1}: } {warning if object is already a non-reactive or
  reactive object or if any references does not exist yet}
  \item{\code{2}: } {error if object is already a non-reactive or
  reactive object or if any references do not exist yet}
}}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Checks prerequisites to determine if its valid to set the reactive object as
desired. This decision is influenced by the level of strictness that is
imposed via \code{strict}.
}
\examples{
\dontrun{

## NOTE
## This function has a strong internal character as it is mainly designed to 
## be used inside of 'setReactiveS3()'.
## Thus not all aspects can be properly illustrated if the function is not 
## used in its default context

## Start with a fresh registry state //
resetRegistry()

suppressWarnings(rm(x_1))
input <- reactr::ReactiveObject.S3(
  id = "x_1",
  value = 10
)

## Error reason: object does not exist yet //
try(checkReactivityPrerequisites(input = input, strict = 2))

## Error reason: object exists but is non-reactive //
x_1 <- 10
try(checkReactivityPrerequisites(input = input, strict = 2))

## Error reason: reactive object already exists //
## This cannot be illustrated outside the context of 'setReactiveS3()'

## All prerequisites met //
checkReactivityPrerequisites(input = input)

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{checkReactivityPrerequisites-ReactiveObject.S3-method},
    \link[reactr]{setReactiveS3}
}
}

