% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reacttools.R
\name{component}
\alias{component}
\title{Create a React component}
\usage{
component(name, varArgs = list())
}
\arguments{
\item{name}{Name of the React component, which must start with an upper-case
character.}

\item{varArgs}{Attributes and children of the element to pass along to
\code{\link[htmltools]{tag}} as \code{varArgs}.}
}
\value{
An htmltools \code{\link[htmltools]{tag}} object
}
\description{
Create a React component
}
\examples{
component("ParentComponent",
  list(
    x = 1,
    y = 2,
    component("ChildComponent"),
    component("OtherChildComponent")
  )
)
}
