% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getReactive}
\alias{getReactive}
\title{Get Reactive Object (generic)}
\usage{
getReactive(id, where = parent.frame(), hidden = FALSE, ...)
}
\arguments{
\item{id}{\strong{Signature argument}.
Object containing name/ID information.}

\item{where}{\strong{Signature argument}.
Object containing location information.}

\item{hidden}{\code{\link{logical}}.
\code{TRUE}: retrieve entire \strong{invisible} object/instance from registry;
\code{FALSE}: retrieve the visible part of the object/instance.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Retrieves an reactive object, or, to be more precise, the \strong{visible}
part of the object which corresponds to the \code{value} field of the
invisible instance of class \code{\link[reactr]{ReactiveObject.S3}}.
}
\examples{
\dontrun{ 

setReactive(id = "x_1", value = 10)
getReactive(id = "x_1")
res <- getReactive(id = "x_1", hidden = TRUE)
res
class(res)

identical(res, getFromRegistry(id = "x_1"))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{getReactive-character-environment-method}
}
}

