% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getRegistry,missing-method}
\alias{getRegistry,missing-method}
\alias{getRegistry-missing-method}
\title{Get Registry (missing)}
\usage{
\S4method{getRegistry}{missing}(where = NULL, ...)
}
\arguments{
\item{where}{\code{\link{missing}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link[reactr]{getRegistry-environment-method}}.
}
\description{
See generic: \code{\link[reactr]{getRegistry}}
}
\examples{
\dontrun{

## Set reactive so that there are elements in registry //  
resetRegistry()
setReactiveS3(id = "x_1", value = Sys.time())  
  
registry <- getRegistry()
registry
ls(registry)
identical(registry, getOption("reactr")$.registry)

## Clean up //
resetRegistry()

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{getRegistry-environment-method},
    \link[reactr]{removeFromRegistryByUid}
}
}

