% Generated by roxygen2 (4.0.2): do not edit by hand
\name{getRegistry}
\alias{getRegistry}
\title{Get Registry (generic)}
\usage{
getRegistry(where = NULL, ...)
}
\arguments{
\item{where}{\strong{Signature argument}.
Object containing registry location information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Convenience function to retrieve the registry located in
\code{getOption("reactr")$.registry}.
}
\examples{
\dontrun{

## Set reactive so that there are elements in registry //  
resetRegistry()
setReactiveS3(id = "x_1", value = Sys.time())  
  
registry <- getRegistry()
registry
ls(registry)
identical(registry, getOption("reactr")$.registry)

## Clean up //
resetRegistry()

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{getRegistry-environment-method},
    \link[reactr]{rmFromRegistryByUid}
}
}

