% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ReactiveObject.S3}
\alias{ReactiveObject.S3}
\title{Class: ReactiveObject.S3}
\usage{
ReactiveObject.S3(.x, .references = character(), id = character(),
  uid = character(), value = NULL, where = parent.frame(),
  checksum = character(), exists_visible = FALSE, has_cached = FALSE,
  is_invalid = FALSE, registry = getRegistry(),
  references_pull = new.env(parent = emptyenv()),
  references_push = new.env(parent = emptyenv()), has_pull_refs = FALSE,
  has_push_refs = FALSE, has_pushed = FALSE, is_running_push = FALSE,
  func = NULL, checksums_ref = new.env(parent = emptyenv()),
  condition = NULL)
}
\arguments{
\item{.x}{\code{\link{ANY}}. An object of an arbitrary class whose class
   attribute should be updated so that it becomes an instance of class
   \code{ReactiveObject.S3}. Mainly intended for rapid prototyping
   purposes}
}
\value{
Instance of class \code{ReactiveObject.S3}.
}
\description{
Class representing the system state (S3) and its constructor function.
}
\details{
Instances of this class are implicitly created when calling
\code{\link[reactr]{setReactiveS3}} and stored in the registry. Objects of
this class can be thought of as the "hidden parts" of the reactive objects
of whom actually only field \code{value} is visible to the user or other
functions consuming the reactive object.
}
\section{Fields}{

\describe{
\item{\code{id}}{\code{\link{character}}.
Object ID.
Initial: \code{character()}.}

\item{\code{uid}}{\code{\link{character}}.
Object ID.
Initial: \code{character()}.
Automatically computed once \code{id} is
specified:
\code{digest::digest(list(id = id, where = capture.output(eval(where))))}.}

\item{\code{value}}{\code{\link{ANY}}.
Actual object value / cached value.
Initial: \code{NULL}.}

\item{\code{where}}{\code{\link{environment}}.
Environment of reactive object.
Initial: \code{parent.frame()}.}

\item{\code{checksum}}{\code{\link{character}}.
Object value checksum.
Initial: \code{character()}.}

\item{\code{exists_visible}}{\code{\link{logical}}.
Field for tracking if the visible object actually exists already or if this
is a mere "empty container" in the registry.
It is set to \code{TRUE} when the visible object is actually set/created
via \code{\link[reactr]{setReactiveS3}}.
Initial: \code{FALSE}.}

\item{\code{has_cached}}{\code{\link{logical}}.
Field for tracking if the object already has a cached value or not.
If \code{FALSE}, the binding function (if there is any) is executed and
after that the field is set to \code{TRUE} to signal that a cached value
exists.
Initial: \code{FALSE}.}

\item{\code{is_invalid}}{\code{\link{logical}}.
Field for propagating the invalidity of referenced objects to its
dependees. It is set to \code{TRUE} when an reactive object is unset or
removed.
Initial: \code{FALSE}.}

\item{\code{registry}}{\code{\link{environment}}.
Reference to the registry environment. Important for retrieving and
     comparing checksum values.
Initial: \code{getRegistry()}.}

\item{\code{references_pull}}{\code{\link{environment}}.
Environment storing information of referenced objects.
Initial: \code{new.env(parent = emptyenv())}.}

\item{\code{func}}{\code{\link{function}}.
Binding function.
Initial: \code{NULL}.}

\item{\code{checksums_ref}}{\code{\link{environment}}.
Environment for caching checksums of referenced objects.
Initial: \code{new.env(parent = emptyenv())}.}

\item{\code{condition}}{\code{\link{condition}} (at least by inheritance).
If a condition has been signaled, this field is assigned a respectiv
condition object that is triggered when \code{.self$value} is requested.
See \code{\link[base]{signalCondition}} and
\code{\link[conditionr]{signalCondition}}
Initial: \code{NULL}.}
}}
\section{Intended use of this class}{


This S3 class, or to be more precise its constructor function, exists mainly
for rapid prototyping purposes.
This is mainly reflected in the fact, that when specifying \code{.x}, this
constructor function will simply update the \code{class} attribute of
whatever object has been provided.

However, it also allows for a more formal OOP-style of rapid
prototyping by offering explicit \emph{class fields} (all arguments except
\code{.x}). Nevertheless, it is probably advisable to switch to an
explicit formal approach such as \emph{S4} and/or \emph{Reference Classes}
once the package or application has reached a certain state of maturity.
}
\examples{
\dontrun{

## Informal use (intended mainly for rapid prototyping) //
## Takes *any* object and simply changes the class attributes
ReactiveObject.S3(
  list(
    id = "x_1",
    value = 10
  )
)  
ReactiveObject.S3(TRUE)  

## Formal use (explicitly using 'fields') //
res <- ReactiveObject.S3()
ls(res)
res <- ReactiveObject.S3(
  id = "x_1",
  value = 10
)
res$id
res$uid
## --> automatically computed; important for handling checksum values

## Recommended: include namespace //
## Regardless if you plan on using this class in an informal or formal way
reactr::ReactiveObject.S3(
  id = "x_1",
  value = 10
)

##------------------------------------------------------------------------------
## Methods //
##------------------------------------------------------------------------------

setReactiveS3(id = "x_1", value = 10)
obj <- reactr::ReactiveObject.S3(
  id = "x_1",
  value = 10
)

## Compute checksum //
digest::digest(x_1)
obj$checksum
obj$value <- x_1 <- 100
digest::digest(x_1)
obj$computeChecksum()

## Compute UID //
x_1_uid <- getObjectUid(id = "x_1")
obj$uid
obj$computeUid()
## --> automatically executed in constructor based on 'obj$id' and 'obj$where'

## Copy //
obj$copy(id = "x_1_copied")
x_1_copied
x_1 <- 100
x_1
x_1_copied
## --> independent
getFromRegistry("x_1")
getFromRegistry("x_1_copied")
## --> independent

## Has references //
obj$hasReferences()

## Register and unregister //
obj$register()
exists(obj$uid, getRegistry())
## --> already registered through call to 'setReactiveS3()'
obj$unregister()
exists(obj$uid, getRegistry())
obj$register()
exists(obj$uid, getRegistry())

## Remove //
## Remove visible object from registry and itself
obj$remove()
try(x_1)
## --> removed 

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{setReactiveS3}
}
}

