% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reacttools.R
\name{reactMarkup}
\alias{reactMarkup}
\title{Prepare data that represents a single-element character vector, a React
component, or an htmltools tag for sending to the client.}
\usage{
reactMarkup(tag)
}
\arguments{
\item{tag}{character vector or React component or
\code{\link[htmltools]{tag}}}
}
\value{
A reactR markup object suitable for being passed to
  \code{\link[htmlwidgets]{createWidget}} as widget instance data.
}
\description{
Tag lists as returned by \code{htmltools tagList} are not currently
supported.
}
