% Generated by roxygen2 (4.0.2): do not edit by hand
\name{resetRegistry}
\alias{resetRegistry}
\title{Reset Registry (generic)}
\usage{
resetRegistry(where = getRegistry(), ...)
}
\arguments{
\item{where}{\strong{Signature argument}.
Object containing registry location information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Resets the registry located in \code{getOption("reactr")$.registry}.
}
\examples{
\dontrun{

## Set reactives so registry has entries //  
setReactive(id = "x_1", value = 10)
setReactive(id = "x_2", value = function() {
  .ref_1 <- get("x_1")
})

## Inspect current state of registry //
registry <- getRegistry()
showRegistry()
## --> two entries corresponding to invisible objects associated to `x_1` 
## and `x_2`

## Reset //
resetRegistry()
showRegistry()
## --> empty 

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{resetRegistry-environment-method},
    \link[reactr]{rmFromRegistryByUid}
}
}

