% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold_widget.R
\name{scaffoldReactWidget}
\alias{scaffoldReactWidget}
\title{Create implementation scaffolding for a React.js-based HTML widget}
\usage{
scaffoldReactWidget(name, npmPkgs = NULL, edit = interactive())
}
\arguments{
\item{name}{Name of widget}

\item{npmPkgs}{Optional \href{https://npmjs.com/}{NPM} packages upon which
this widget is based which will be used to populate \code{package.json}.
Should be a named list of names to
\href{https://docs.npmjs.com/files/package.json#dependencies}{versions}.}

\item{edit}{Automatically open the widget's JavaScript source file after
creating the scaffolding.}
}
\description{
Add the minimal code required to implement a React.js-based HTML widget to an
R package.
}
\note{
This function must be executed from the root directory of the package
  you wish to add the widget to.
}
