% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{showRegistry,environment-method}
\alias{showRegistry,environment-method}
\alias{showRegistry-environment-method}
\title{Get Hash Registry (environment)}
\usage{
\S4method{showRegistry}{environment}(where = NULL, ...)
}
\arguments{
\item{where}{\code{\link{environment}}.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
\code{\link{logical}}. \code{TRUE}: successful; \code{FALSE}: fail.
}
\description{
See generic: \code{\link[reactr]{showRegistry}}
}
\examples{
\dontrun{

showRegistry()

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   \link[reactr]{showRegistry},
    \link[reactr]{getRegistry}
}
}

