% Generated by roxygen2 (4.0.2): do not edit by hand
\name{showRegistry}
\alias{showRegistry}
\title{Show Registry Content (generic)}
\usage{
showRegistry(where = NULL, ...)
}
\arguments{
\item{where}{\strong{Signature argument}.
Object containing registry location information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Convenience function to show the content of the registry in
\code{getOption("reactr")$.registry}.
}
\examples{
\dontrun{

showRegistry()

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{showRegistry-environment-method},
    \link[reactr]{getRegistry}
}
}

