% Generated by roxygen2 (4.0.2): do not edit by hand
\name{unsetReactive}
\alias{unsetReactive}
\title{Unset Reactive Object (generic)}
\usage{
unsetReactive(id, where = parent.frame(), ...)
}
\arguments{
\item{id}{\strong{Signature argument}.
Object containing path-like ID information.}

\item{where}{\strong{Signature argument}.
Object containing location information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\description{
Removes the reactive \strong{character} from an object, i.e. its
binding(s) to other objects and the ability for other objects to have
bindings to it. It is equivalent to transforming the object to one that
has been assigned via \code{\link[base]{assign}} or \code{\link[base]{<-}}
instead of \code{\link[reactr]{setReactiveS3}} and thus in turn
by \code{\link[base]{makeActiveBinding}}.

Note that it is \strong{not} equivalent to removing/deleting the object!
See \code{\link[reactr]{removeReactive}} for this purpose.
}
\note{
The main S4 method is
\code{\link[reactr]{unsetReactive-character-environment-method}}.
}
\section{Implications with respect to observing variables}{


If other reactive variables have been observing the reactive variable that
has been unset, from this point on they will simply return the last value
that has been cached. \strong{So there is no actual reactive binding anymore}.
}
\examples{
\dontrun{

## Start with a fresh hash registry //
resetRegistry()

setReactiveS3(id = "x_1", value = 10)
setReactiveS3(id = "x_2", 
  value = function(refs = list(x_1 = parent.frame())) x_1 * 2
)

## Illustrate reactiveness //
x_1
x_2
x_1 <- 50
x_1 
x_2

## Unset reactive --> turn it into a regular object again //
unsetReactive(id = "x_1", where = where)

## Illustrate removed reactiveness //
x_1
x_2
x_1 <- 10
x_1
x_2
## --> 'x_1' is not a reactive object anymore; from now on, 'x_2' simply returns
## the last value that has been cached

## Implications with respect to hash registry //
## Entry corresponding to 'x_1' set in 'environment()' has been removed
exists(getObjectUid(id = "x_1", where = where, getRegistry()))
## [1] FALSE

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
  	\link[reactr]{unsetReactive-character-environment-method}
}
}

