% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{unsetReactiveByUid,character-method}
\alias{unsetReactiveByUid,character-method}
\alias{unsetReactiveByUid-method_main}
\title{Unset Reactive Object (character)}
\usage{
\S4method{unsetReactiveByUid}{character}(uid, ...)
}
\arguments{
\item{uid}{\strong{Signature argument}.
Object containing UID information.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}

\item{id}{\code{\link{character}}.}
}
\value{
\code{\link{logical}}.
   \code{TRUE}: successfully unset;
   \code{FALSE}: failed to unset.
}
\description{
See generic: \code{\link[reactr]{unsetReactiveByUid}}
}
\examples{
\dontrun{

setReactiveS3(id = "x_1", value = 10)
unsetReactiveByUid(uid = getObjectUid("x_1"))
isReactive("x_1")

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/reactr}
}
\seealso{
\code{
   Generic: \link[reactr]{unsetReactiveByUid},
    \link[reactr]{unsetReactive}
}
}

